restructured into Menus, MenuComponents, and Utils
